Add files via upload
The aim of this research is to identify the most reliable drag correlation for
computational modelling of dispersed Multiphase flows in a 2D Fluidized bed. The present
study employs the open-source CFD toolbox, OpenFOAM, to investigate the effect of various
drag correlations in the literature on the Mean Eulerian particle velocity distribution of the gassolid fluidized bed. Superficial gas velocities of 2.19 m/s and 3.28 m/s are considered at the
inlet for the transient multiphase simulation. The time-averaged particle velocity is estimated
at five lateral locations of the fluidized bed for validation against the experimental data of the
National Energy Technology Laboratory (NETL). The geometry and boundaries of the 2D
fluidized bed used for this study are shown in Figure 1 along with few important parameters in
Table 1. Finally, the particle velocity distribution is compared between a drag correlation
proposed by Syamlal O’Brien and two combined drag correlations of Gidaspow